+++ babel.config.js
@@ -1,3 +1,13 @@
 module.exports = {
   presets: ['@babel/preset-env', '@babel/preset-react'],
+  plugins: [
+    [
+      'babel-plugin-styled-components',
+      {
+        ssr: true,
+        displayName: true,
+        fileName: false,
+      },
+    ],
+  ],
 }