+++ packages/card/src/index.test.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { mount, shallow } from 'enzyme'
-
-import Button from '../../button'
+import Button from '@natterstefan/ns-button-js'
 
 import Card from '.'